Multi Agent Researcher Google ADK Project/pyproject.toml
[project]
name = "google-adk-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.22.1",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
